restructuring of service creation
