+++ config/python/requirements.txt
@@ -20,3 +20,4 @@ requests==2.27.1
 six==1.16.0
 urllib3==1.26.9
 wcwidth==0.2.5
+pyyaml==6.0
