+++ libs/json/templates/K8s-SERVICE-BINDING.json
@@ -1,0 +1,13 @@
+{
+    "apiVersion": "services.cloud.sap.com/v1",
+    "kind": "ServiceBinding",
+    "metadata": {
+        "name": "my-service-binding"
+    },
+    "spec": {
+        "serviceInstanceName": "my-service-instance",
+        "externalName": "my-service-instance-external",
+        "secretName": "mySecret",
+        "parameters": {}
+    }
+}
