+++ libs/python/helperEnvBTP.py
@@ -1,0 +1,10 @@
+def get_btp_service_status(btpUsecase, service):
+    return "status"
+
+
+def create_btp_service(btpUsecase, service):
+    return "service"
+
+
+def getStatusResponseFromCreatedBTPInstance(btpUsecase, instancename):
+    return "status"
